fix(service): service restart handling with watch requisite
Previously, the state for the keepalived service included the enable
attribute. This would prevent the watch requisite to restart the service
if the service was already running but not enabled.

For reference, https://docs.saltproject.io/en/latest/ref/states/requisites.html#watch

Split service handling in service.enabled and service.running.